requirements.txt
streamlit
requests
beautifulsoup4
supabase
python-dotenv
pandas
